agent/requirements.txt
docker==7.1.0
httpx==0.27.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
schedule==1.2.1
netifaces==0.11.0
psutil==5.9.8
